refactor: ScheduleProvider와 ScheduleTables의 import 경로를 업데이트하여 코드 구조 개선
